add logging, improve exceptions handling
add logging using logback

+++ src/main/java/pl/bzpb/bookforum/services/BookService.java
@@ -1,10 +1,13 @@
 package pl.bzpb.bookforum.services;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import pl.bzpb.bookforum.dao.BookRepo;
 import pl.bzpb.bookforum.dao.entity.Book;
 import pl.bzpb.bookforum.dao.entity.Rating;
+import pl.bzpb.bookforum.services.exceptions.BookAlreadyExist;
 
 import java.util.Collections;
 import java.util.List;
@@ -14,6 +17,7 @@ import java.util.Optional;
 @Service
 public class BookService {
 
+    private static final Logger log = LoggerFactory.getLogger(BookService.class);
     BookRepo bookRepo;
 
     @Autowired
@@ -21,44 +25,50 @@ public class BookService {
         this.bookRepo = bookRepo;
     }
 
-    public void addBook(Book book) {
+    public void addBook(Book book) throws BookAlreadyExist {
         Optional<Book> optionalBook = bookRepo.findById(book.getIsbn());
 
         if (optionalBook.isPresent()) {
-            //wyrzucic wyjatek throw BookAlreadyExist
+            throw new BookAlreadyExist();
         }
         bookRepo.save(book);
+        log.info("Book: {} has been added", book.getIsbn());
     }
 
-    public List<Book> getAllBooks () throws NoSuchElementException{
-        try{
-            return (List<Book>) bookRepo.findAll();
-        } catch (Exception e){
-            throw  new NoSuchElementException();
-        }
+    public List<Book> getAllBooks () {
+        log.info("List of books has been returned");
+        return (List<Book>) bookRepo.findAll();
     }
 
     public void deleteBook(Long id) throws Exception {
-        try{
-            Book book = bookRepo.findById(id).get();
-            bookRepo.delete(book);
-        } catch (Exception e){
-            throw new Exception();
+        Optional<Book> optionalBook = bookRepo.findById(id);
+
+        if(optionalBook.isEmpty()){
+            throw new NoSuchElementException();
         }
+
+        Book book = optionalBook.get();
+        bookRepo.delete(book);
+        log.info("Book: {} has been deleted", book.getIsbn());
+
     }
 
-    public void editBook(Book bookedit, Long id) throws Exception {
-        try{
-            Book book = bookRepo.findById(id).get();
-            if(!bookedit.getAuthor().equals("")){
-                book.setAuthor(bookedit.getAuthor());
-            }
-            if(!bookedit.getTitle().equals("")){
-                book.setTitle(bookedit.getTitle());
-            }
-            bookRepo.save(book);
-        } catch (Exception e){
-            throw new Exception();
+    public void editBook(Book bookToUpdate, Long id) throws NoSuchElementException {
+        Optional<Book> optionalBook = bookRepo.findById(id);
+
+        if(optionalBook.isEmpty()){
+            throw new NoSuchElementException();
         }
+
+        Book book = optionalBook.get();
+
+        if(!bookToUpdate.getAuthor().equals("")){
+            book.setAuthor(bookToUpdate.getAuthor());
+        }
+        if(!bookToUpdate.getTitle().equals("")){
+            book.setTitle(bookToUpdate.getTitle());
+        }
+        bookRepo.save(book);
+        log.info("Book: {} has been changed", book.getIsbn());
     }
 }
